make test propagator private for testing
