+++ internal/internal_workflow_client_test.go
@@ -62,6 +62,72 @@ type (
 	}
 )
 
+// stringMapPropagator propagates the list of keys across a workflow,
+// interpreting the payloads as strings.
+type stringMapPropagator struct {
+	keys map[string]struct{}
+}
+
+// NewStringMapPropagator returns a context propagator that propagates a set of
+// string key-value pairs across a workflow
+func NewStringMapPropagator(keys []string) ContextPropagator {
+	keyMap := make(map[string]struct{}, len(keys))
+	for _, key := range keys {
+		keyMap[key] = struct{}{}
+	}
+	return &stringMapPropagator{keyMap}
+}
+
+// Inject injects values from context into headers for propagation
+func (s *stringMapPropagator) Inject(ctx context.Context, writer HeaderWriter) error {
+	for key := range s.keys {
+		value, ok := ctx.Value(contextKey(key)).(string)
+		if !ok {
+			return fmt.Errorf("unable to extract key from context %v", key)
+		}
+		writer.Set(key, []byte(value))
+	}
+	return nil
+}
+
+// InjectFromWorkflow injects values from context into headers for propagation
+func (s *stringMapPropagator) InjectFromWorkflow(ctx Context, writer HeaderWriter) error {
+	for key := range s.keys {
+		value, ok := ctx.Value(contextKey(key)).(string)
+		if !ok {
+			return fmt.Errorf("unable to extract key from context %v", key)
+		}
+		writer.Set(key, []byte(value))
+	}
+	return nil
+}
+
+// Extract extracts values from headers and puts them into context
+func (s *stringMapPropagator) Extract(ctx context.Context, reader HeaderReader) (context.Context, error) {
+	if err := reader.ForEachKey(func(key string, value []byte) error {
+		if _, ok := s.keys[key]; ok {
+			ctx = context.WithValue(ctx, contextKey(key), string(value))
+		}
+		return nil
+	}); err != nil {
+		return nil, err
+	}
+	return ctx, nil
+}
+
+// ExtractToWorkflow extracts values from headers and puts them into context
+func (s *stringMapPropagator) ExtractToWorkflow(ctx Context, reader HeaderReader) (Context, error) {
+	if err := reader.ForEachKey(func(key string, value []byte) error {
+		if _, ok := s.keys[key]; ok {
+			ctx = WithValue(ctx, contextKey(key), string(value))
+		}
+		return nil
+	}); err != nil {
+		return nil, err
+	}
+	return ctx, nil
+}
+
 func TestHistoryEventIteratorSuite(t *testing.T) {
 	s := new(historyEventIteratorSuite)
 	suite.Run(t, s)
